add more load tests and fixtures
